+++ src/main/java/com/ssap/SSAPIDE/domain/member/User.java
@@ -1,5 +1,6 @@
 package com.ssap.SSAPIDE.domain.member;
 
+import jakarta.validation.constraints.NotBlank;
 import lombok.*;
 
 @Getter
@@ -9,7 +10,12 @@ import lombok.*;
 public class User {
     private Long id; // memory repository 이용 필드
 
+    @NotBlank
     private String email;
+
+    @NotBlank
     private String password;
+
+    @NotBlank
     private String name;
 }
