+++ src/main/java/com/ssap/SSAPIDE/dto/ErrorField.java
@@ -1,0 +1,13 @@
+package com.ssap.SSAPIDE.dto;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.RequiredArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@RequiredArgsConstructor
+public class ErrorField {
+    private final String field;
+}
