+++ src/main/java/com/ssap/SSAPIDE/dto/SignupRequest.java
@@ -1,5 +1,6 @@
 package com.ssap.SSAPIDE.dto;
 
+import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
@@ -8,8 +9,16 @@ import lombok.ToString;
 @Setter
 @ToString
 public class SignupRequest {
+
+    @NotBlank
     private String email;
+
+    @NotBlank
     private String password;
+
+    @NotBlank
     private String passwordConfirm;
+
+    @NotBlank
     private String name;
 }
